feat: databaseUtils
